requirements.txt
notion-client==2.2.1
python-dotenv==1.0.0
boto3==1.34.0
astrapy==0.7.4 